==> R_scripts/02_rate_generation_all_sets.Rmd <==
---
title: "Generating rates of change and CIs for all SETs"
author: "Kim Cressman"
date: "`r Sys.Date()`"
output:
  word_document:
    toc: true
    toc_depth: 2
  pdf_document:
    toc: yes
  html_document:
    toc: yes
    toc_float: yes
always_allow_html: yes
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message = FALSE, warning = FALSE}

library(dplyr)
library(tidyr)
library(purrr)
library(readxl)
library(stringr)
library(lubridate)
library(here)
library(janitor)
library(nlme)
library(ggplot2)
library(forcats)
library(flextable)
```

Read in combined data file.  

```{r}
in_path <- here::here("data", "processed_combined", "all_sites.csv")
dat <- read.csv(in_path, stringsAsFactors = FALSE) %>% 
    mutate(date = lubridate::ymd(paste(year, month, day, sep = "-")),
           date2 = lubridate::decimal_date(date))
```



***

Read in metadata.  

```{r}
in_path <- here::here("metadata", "combined", "all_sites_metadata.csv")
mdat <- read.csv(in_path, stringsAsFactors = FALSE)
```

Make sure all SETs we have in the `dat` data frame are present in the metadata, and that they have lat/long coordinates. Get rid of anything in the metadata that's not in the data (some reserves have information on SETs that aren't up and running yet).  

```{r}
# first pull out set_id from both data frames
data_setid <- unique(as.character(dat$set_id))
metadata_setid <- unique(mdat$set_id)
# find set_ids that are in the data, but not in the metadata
dat_not_m <- setdiff(data_setid, metadata_setid)
# find set_ids that are in the metadata, but not in the data
m_not_dat <- setdiff(metadata_setid, data_setid)

# print a message
if (length(dat_not_m) > 0) {
    toprint <- paste(dat_not_m, collapse = ", ")
    warning(paste0("The following SET IDs exist in your data, but not in your metadata: ", toprint))
}
```

As of 7/2:  

+  APA Little St. Marks was only measured once, in 2016. Will leave it in data file but remove it from data frame for this analysis.  
 


```{r}
if (length(m_not_dat) > 0) {
    toprint <- paste(m_not_dat, collapse = ", ")
    warning(paste0("The following SET IDs exist in your metadata, but not in your data, and will be removed within the current analysis: ", toprint))
}
if (length(dat_not_m) + length(m_not_dat) == 0) {
    print("SET IDs match in your data and metadata files.")
}
```


Weed out the APA station from the data; remove stations in metadata that aren't in data.  

```{r}
to_remove_dat <- c("Little St. Marks")
dat <- filter(dat, !(set_id %in% to_remove_dat))
mdat <- filter(mdat, !(set_id %in% m_not_dat))
```

Check for any stations in the metadata that don't have lat/long coordinates.  

```{r}
sum(is.na(mdat$lat))
mdat %>% 
  filter(is.na(lat)) %>% 
  select(reserve, set_id)
```

### Come back to this later if there are problems. Might need to just remove those from the data frame and treat them like the stations for which we have no metadata.  

***

# Filter by time and number of samples  

Filter to only analyze sites with at least 5 measurements over at least 4.5 years.  

```{r}
to_keep <- dat %>% 
       group_by(reserve, set_id, arm_position, pin_number) %>% 
        summarize(count = n(),
                  daterange = max(date2) - min(date2)) %>% 
        filter(count >= 5,
               daterange >= 4.5)


to_analyze <- dat %>% 
        filter(set_id %in% unique(to_keep$set_id))
```

Working with how many SETs?  

```{r}
length(unique(to_analyze$set_id))
```

At how many reserves?  

```{r}
length(unique(to_analyze$reserve))
```

# Generate rates  

Run random-intercept LMMs as in the main project: fixed effect = date; random effects = pin nested within arm.  

```{r}
models2 <- to_analyze %>%
    group_by(reserve, set_id) %>%
    do(mod = lme(pin_height ~ date, data = ., random = ~1|arm_position/pin_number, na.action = na.omit))
```

Pull out information from the models.  

```{r}
lmm_out <- models2 %>% 
  mutate(rate = intervals(mod, which = "fixed")$fixed["date", 2] * 365.25,
         CI_low = intervals(mod, which = "fixed")$fixed["date", 1] * 365.25,
         CI_high = intervals(mod, which = "fixed")$fixed["date", 3] * 365.25) %>% 
  select(-mod)
```

```{r}
lmm_out %>% 
  flextable() %>% 
  theme_booktabs() %>% 
  autofit()
```

***
***

# Join rates to metadata  

Need to do a full join; there are some sites in the data that aren't in metadata; and some sites in metadata that didn't have enough data to be analyzed but we can put them on a map with some icon.  

```{r}
all_out <- full_join(mdat, lmm_out, by = c("reserve", "set_id"))
```

Write that out to a file.  

```{r}
write.csv(all_out, here::here("data", "intermediate", "all_rates_and_metadata.csv"))
```


Get rid of any sites that are missing *both* a rate and a location.  

```{r}
# find 'em
ditch <- all_out %>% 
  filter(is.na(lat),
         is.na(rate)) %>% 
  select(reserve, set_id)

# ditch 'em
all_out <- all_out %>% 
  filter(!(set_id %in% ditch$set_id))
```


Read in SLR rates and join them to the data frame so we can make comparisons.  

```{r}
slr_path <- here::here("metadata", "slr_rates.csv")
slr_rates <- read.csv(slr_path, stringsAsFactors = FALSE) %>% 
  clean_names() %>% 
  remove_empty(which = c("rows", "cols"))

to_join <- slr_rates %>% 
  mutate(slr_CI_low = slr_rate_mm_yr - x_95_ci,
         slr_CI_high = slr_rate_mm_yr + x_95_ci,
         yr19_CI_low = slr_19yr - ci95_19yr,
         yr19_CI_high = slr_19yr + ci95_19yr) %>% 
  select(reserve, 
         slr_rate = slr_rate_mm_yr,
         slr_CI_low,
         slr_CI_high,
         slr_19yr,
         yr19_CI_low,
         yr19_CI_high,
         slr_start = data_start,
         yr19_start,
         slr_end = data_end)

all_to_compare <- left_join(all_out, to_join, by = "reserve") 
```

Weed out some columns; generate comparison columns for both 0 and slr.  

```{r}
rates_comp <- all_to_compare %>% 
  select(reserve, 
         set_id, 
         user_friendly_set_name, 
         numerical_order,
         lat, long, 
         rate, CI_low, CI_high,
         slr_rate, slr_CI_low, slr_CI_high,
         slr_19yr, yr19_CI_low, yr19_CI_high) %>% 
  mutate(dir_0 = case_when(CI_high < 0 ~ "dec_sig",
                           CI_low > 0  ~ "inc_sig",
                           rate < 0 ~ "dec_nonsig",
                           rate > 0 ~ "inc_nonsig",
                           TRUE ~ "not_enough_info"),
         dir_slr = case_when(CI_high < slr_CI_low ~ "dec_sig",
                             CI_low > slr_CI_high  ~ "inc_sig",
                             rate < slr_rate ~ "dec_nonsig",
                             rate > slr_rate ~ "inc_nonsig",
                             TRUE ~ "not_enough_info"),
         dir_19yr = case_when(CI_high < yr19_CI_low ~ "dec_sig",
                              CI_low > yr19_CI_high  ~ "inc_sig",
                              rate < slr_19yr ~ "dec_nonsig",
                              rate > slr_19yr ~ "inc_nonsig",
                              TRUE ~ "not_enough_info"),
         set_to_slr = rate / slr_rate,
         set_to_19yr = rate / slr_19yr) 
```


Summarize the categories.  

```{r}
rates_comp_summary <- rates_comp %>% 
  select(reserve, set_id, dir_0, dir_slr, dir_19yr) %>% 
  gather(key = comparison, value = value, -reserve, -set_id) %>%
  group_by(reserve, comparison, value) %>% 
  summarize(total = n()) %>% 
  mutate(value = factor(value, levels = c( 
                             "dec_sig", "dec_nonsig", 
                             "not_enough_info", 
                             "inc_nonsig", "inc_sig"))) 

```


```{r}
# set up color palette; associate it with the correct factor levels
# from colorbrewer2.org, 5-class RdYlBu, substituting a gray for the yellow middle value:
# to_color <- c("#d7191c", "#fdae61", "#f0f0f0", "#abd9e9", "#2c7bb6")


# from reserve_template; these are the colors used
# map_pal <- c("#c00000", "#2f5597", "#7f7f7f", "#fffacd")
# okay the yellow and endpoints are the same below
# this script also needs some intermediate colors; so it all matches. good.


# dark colors consistent with map icons; mid red/blue from color brewer 5-class RdBu:
to_color_lightgray <- c("#c00000", "#f4a582", "#f0f0f0", "#92c5de", "#2f5597")  
to_color <- c("#c00000", "#f4a582", "#fffacd", "#92c5de", "#2f5597") # using yellow for contrast on map 
names(to_color) <- levels(rates_comp_summary$value)
to_label <- c("Lower, CIs don't overlap", "Lower, CIs overlap", 
                     "Not enough information",
                     "Higher, CIs overlap", "Higher, CIs don't overlap")


# generate some info to calculate y positions
to_plot <- rates_comp_summary %>% 
  ungroup() %>% 
  arrange(reserve, comparison, value) %>% 
  group_by(reserve, comparison) %>%
  arrange(reserve, comparison, value) %>% 
  mutate(total_cumu = cumsum(total),
         ypos = sum(total) - (total_cumu - total/2))



# going to need to do a nested loop
# for each reserve...
#  .... for each comparison type

reserves <- unique(to_plot$reserve)
comp_type <- unique(to_plot$comparison)


for(i in seq_along(reserves)){
  # pull out the reserve
  dat_sub <- to_plot %>% 
    filter(reserve == reserves[i])
  
  # loop through comparison type
  for(j in seq_along(comp_type)){
    to_plot_sub <- dat_sub %>% 
      filter(comparison == comp_type[j])
    
    # set up names
    res_to_plot <- reserves[i]
    comp_to_plot <- comp_type[j]
    nsets <- sum(to_plot_sub$total)
    
    # to use on maps:
    p <- ggplot(to_plot_sub) +
      geom_col(aes(x = "", y = total, fill = value), color = "black", size = 1) +
      scale_fill_manual(values = to_color,
                        labels = to_label,
                        drop = FALSE) +
      coord_polar("y", start = 0) +
      theme_void() +   
      theme(legend.position = "none")
    
    # to use individually:
    q <- p +
      geom_text(aes(x = "", y = ypos, label = total, fontface = "bold"),
                color = "gray10",
                size = 5) +
      labs(title = res_to_plot,
          subtitle = paste0("(n = ", nsets, ")")) +
      theme(plot.title = element_text(hjust = 0.5, vjust = -3),
             plot.subtitle = element_text(hjust = 0.5, vjust = -3),
            legend.position = "right")  
    
    
    
    # save the plots
    file_out <- paste0(res_to_plot, "_", comp_to_plot, ".svg")
    out_path <- here::here("R_output", "figures", "for_maps", file_out)
    ggsave(out_path, plot = p, bg = "transparent",
           width = 1, height = 1, units = "in")
    
    file_out <- paste0(res_to_plot, "_", comp_to_plot, "_full.svg")
    out_path <- here::here("R_output", "figures", "for_individual_use", file_out)
    ggsave(out_path, plot = q, bg = "transparent",
           width = 7, height = 7, units = "in")
  }
}

```

Those pie charts can be used to represent each reserve. Need to generate a data frame with a single point per reserve. I'll just average all the lats and longs for each one.  

```{r}
locations <- rates_comp %>% 
  group_by(reserve) %>% 
  summarize(lat = mean(lat, na.rm = TRUE),
            long = mean(long, na.rm = TRUE)) %>% 
  filter(!is.na(lat),
         !is.na(long))
```

Write out the files to use in mapping script.  

```{r}
out_path <- here::here("data", "intermediate", "locations.csv")
write.csv(locations, out_path, row.names = FALSE)

out_path <- here::here("data", "intermediate", "rate_summary.csv")
write.csv(all_out, out_path, row.names = FALSE)

out_path <- here::here("data", "intermediate", "rate_comparisons.csv")
write.csv(rates_comp, out_path, row.names = FALSE)

out_path <- here::here("data", "intermediate", "rate_comparisons_summary.csv")
write.csv(rates_comp_summary, out_path, row.names = FALSE)
```





# Graphs  

### To make sure linear models make sense in the first place  

Spoiler alert: some don't.  

```{r, fig.width = 8, fig.height = 11, message = FALSE, warning = FALSE}

ind_reserves <- unique(dat$reserve)

for(i in seq_along(ind_reserves)){
        to_plot <- filter(dat, reserve == ind_reserves[i])
        p <- ggplot(to_plot) +
                geom_point(aes(x = date,
                               y = pin_height,
                               col = as.factor(arm_position)),
                           alpha = 0.5) +
                geom_smooth(aes(x = date, y = pin_height),
                            method = "lm",
                            se = FALSE) +
                facet_wrap(~ set_id, ncol = 4, scales = "free") +
                labs(title = ind_reserves[i],
                     color = "Arm Position") +
                theme_bw()
        print(p)
}

```
